+++ src/main/java/wooteco/subway/path/domain/fare/Fare.java
@@ -1,4 +1,4 @@
-package wooteco.subway.path.domain;
+package wooteco.subway.path.domain.fare;
 
 import wooteco.subway.exception.application.ValidationFailureException;
